requirements.txt
faiss-cpu==1.7.4
numpy==1.21.6
PyYAML==6.0.1